ahora si

+++ src/ProtocolNew.cpp
@@ -10443,14 +10443,14 @@ ConsoleMsg::ConsoleMsg() : ServerPacket(ServerPacketID_ConsoleMsg /* 24 */) {
 ConsoleMsg::ConsoleMsg(clsByteQueue* buffer) : ServerPacket(ServerPacketID_ConsoleMsg /* 24 */) {
     buffer->ReadByte(); /* PacketID */
     Chat = buffer->ReadUnicodeString();
-    FontIndex = buffer->ReadInteger();
+    FontIndex = buffer->ReadByte();
 
 }
 
 void ConsoleMsg::serialize(clsByteQueue* buffer) const {
     buffer->WriteByte(ServerPacketID_ConsoleMsg); /* PacketID: 24 */
     buffer->WriteUnicodeString(Chat);
-    buffer->WriteInteger(FontIndex);
+    buffer->WriteByte(FontIndex);
 
 }
